Fix project structure with right service moduling
