+++ workshopCode/src/DomainLayer/ExternalSystems/ProductSupply.java
@@ -1,0 +1,11 @@
+package DomainLayer.ExternalSystems;
+
+import DomainLayer.Models.Purchase;
+import DomainLayer.Models.User;
+
+public class ProductSupply {
+
+    public void supply(Purchase purchase, User user){
+
+    }
+}
