+++ workshopCode/src/DomainLayer/Models/Basket.java
@@ -1,4 +1,8 @@
-package DomainLayer;
+package DomainLayer.Models;
+
+import DomainLayer.Models.Product;
+import DomainLayer.Models.Store;
+import DomainLayer.ProductItem;
 
 import java.util.ArrayList;
 import java.util.List;
