--- workshopCode/src/DomainLayer/ProductSupply.java
@@ -1,8 +1,0 @@
-package DomainLayer;
-
-public class ProductSupply {
-
-    public void supply(Purchase purchase, User user){
-
-    }
-}
