--- workshopCode/src/ServiceLayer/AddToShoppingBasket.java
@@ -1,19 +1,0 @@
-package ServiceLayer;
-
-import DomainLayer.SystemHandler;
-import DomainLayer.SystemLogger;
-
-public class AddToShoppingBasket {
-
-    public String AddToShoppingBasket(String storeName, String productName, int amount){
-        SystemLogger.getInstance().writeEvent("Add to shopping basket command: store name - " + storeName + ", product name - " + productName);
-        try {
-            SystemHandler.getInstance().addToShoppingBasket(storeName, productName, amount);
-            return "Items have been added to basket";
-        }
-        catch (Exception e){
-            SystemLogger.getInstance().writeError("Add to shopping basket error: " + e.getMessage());
-            return e.getMessage();
-        }
-    }
-}
