--- workshopCode/src/ServiceLayer/ViewCart.java
@@ -1,9 +1,0 @@
-package ServiceLayer;
-
-import DomainLayer.SystemHandler;
-
-public class ViewCart {
-    public String execute(){
-        return SystemHandler.getInstance().viewSoppingCart();
-    }
-}
